+++ src/main/java/kitchenpos/tablegroup/domain/TableGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.tablegroup.domain;
 
 import java.time.LocalDateTime;
 import java.util.List;
