+++ src/main/java/kitchenpos/tablegroup/domain/infra/JdbcTemplateTableGroupDao.java
@@ -1,4 +1,4 @@
-package kitchenpos.dao.infrastructure;
+package kitchenpos.tablegroup.domain.infra;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -6,7 +6,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 import javax.sql.DataSource;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.tablegroup.domain.TableGroup;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
