+++ src/test/java/kitchenpos/Fixture.java
@@ -2,7 +2,6 @@ package kitchenpos;
 
 import java.math.BigDecimal;
 import java.util.List;
-import kitchenpos.domain.TableGroup;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menugroup.domain.MenuGroup;
@@ -11,6 +10,7 @@ import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.product.domain.Product;
+import kitchenpos.tablegroup.domain.TableGroup;
 
 @SuppressWarnings("NonAsciiCharacters")
 public class Fixture {
